avio: Add an option 'rw_timeout'
If set non-zero, this limits duration of the retry_transfer_wrapper()
loop, thus affecting ffurl_read*(), ffurl_write(). As soon as
one single byte is successfully received/transmitted, the timer
restarts.

This has further changes by Michael Niedermayer and Martin Storsjö.
---
Andrey's original version didn't reset the timer on successful
transfers, but I think this is closer to how most people may
want it to work.

I also added an AVOption for setting this, to allow it to be set
generically for any URLProtocol.