rtspenc: Add parse_rtcp and rtcp_log options
Ideally, parse_rtcp could be a flag in the rtsp_flags field, but
it's kept as a separate option to be similar to the options in the
RTP muxer.

This handles marshalling of the option into the internal RTP muxers,
and propagating the log back to the RTSP muxer so that the caller
can get it.