configure: Avoid requiring c99wrap for working around msys path issues
The only paths that actually are an issue are absolute unix-style
paths, and the only place they're used with the output arguments
(-Fo, -Fe, -Fi, -out:) are for the temp files within configure.

By setting the default TMPDIR to . for msvc/icl builds, the msys
path conversion isn't an issue any longer, and we can use the
proper form of the file output parameters. This avoids requiring
the c99wrap wrapper for remapping the parameters for cases where
the c99 converter isn't invoked at all.