movenc: Expose start_dts/start_cts (per stream) via AVOptions
This allows writing a moov with an edit list even before the first
packet has been written.

This allows doing random-access writing of fragments when the
start dts of the stream isn't zero, so that the edit list in the moov
is written based on timestamps from the nominal start time set via
an option, while the first packet written corresponds to a later
fragment.

It would probably make more sense to change it to start_pts (which
the caller probably can provide a bit more easily) and adjust the
internals to use that instead of start_cts?

This is an initial step, which only works after avformat_write_header
has been called (because st->priv_data is only initialized at that
point). Options can't be set via a dictionary, and options can't
be set before calling write_header.

TODO: Could this be hooked to the main hierarchy via the AVClass
child_next and child_class_next fields? What would the pros and
cons of that be?

Would there need to be a separate AVClass for all mov muxer
variants' MOVTracks?