omxdec: Add zerocopy output in the decoder
In this case, the decoder doesn't use the get_buffer callback, since
it can't use any specific data layout. Instead a separate AVBufferRef
is allocated and passed to the decoder to fill, and then detached
from the codec and attached to an AVFrame when returned to the caller.

TODO: Add an option for controlling the use of it? Not all codecs
might support OMX_UseBuffer.