movenc: Interpret zero sized packets as a hint for the next pts/dts
Previously, zero sized packets can be used to signal the nominal
start time for a track, when muxing discontinuously. (That is,
to indicate that the track starts at dts=-1, pts=0, while starting
to mux from dts=99,pts=100, for muxing individual fragments.)

After this patch, such zero sized packets can also be used to
indicate the next timestamp in a stream.

For the last packet in a fragment, the muxer uses either pkt->duration
or a heuristic guess based on older packet intervals, to guess what
the start timestamp of the next fragment will be. Once that fragment
has been written, the next fragment must start at that time, and thus
the timestamp of the next packet will be readjusted accordingly.

If the muxer can cut fragments automatically, it will already know
the next packet and can write the proper packet duration at the end
of the fragment (for the one stream that caused the fragment to be cut).

If the caller wants to manually flush fragments by calling
av_write_frame(ctx, NULL), the muxer doesn't know the timestamps of
the next packet, and must resort to guessing.

Such timestamp guessing and readjusting doesn't work perfectly for VFR
streams, even less so for VFR streams with b-frames.

Therefore, allow a caller that uses manual fragment flushing to
signal the duration of the last packet of a fragment (for any streams
where this info is available), by muxing a zero sized packet with the
same timestamp before flushing.