libavformat: Add new public API for interacting with the interleaving queue
This doesn't work for the few muxers that implement custom
interleave_packet functions (gxfenc, mxfenc). It seems to me that
they also could be covered, but that would require changing the
signature of the interleave_packet function to split adding packets
and polling/getting packets, which probably would require a major bump.

---
TODO: apichanges, version bump.