avio: Allow custom IO users to get output bytestream positions for keyframes
This allows callers with avio write callbacks to get the bytestream
positions that correspond to keyframes, suitable for live streaming.

TODO: Is the ABI structure too rigid? An array of struct can't be
enlarged without breaking ABI. A fixed length array (with a couple
entries) should be enough for most practical cases, but having the
caller allocate it allows it to be longer in case that ever would
be useful.

Or should we allow the struct to grow, by having callers
allocate av_stream_pos_struct_size() * nb_output_pos entries,
and having an accessor to read it?

Do we want to keep time? Is it useful to allow marking non-keyframes?
What about streams with multiple video tracks, how is keyframes
tagged there? (Is it worth making an issue of that, given how
much common streaming issues this helps, for the common cases?)